perf(view): NavigationBarが読み込まれる度に認証情報の取得が行われないようにする
